Extracts mocks used in CogitoContract tests to separate files
