+++ workspaces/demo-app/src/test-helpers/index.js
@@ -1,3 +1,6 @@
 export { inRouter, renderInRouter } from './router'
 export { EventWaiter } from './EventWaiter'
 export { GanacheTestNetwork } from './GanacheTestNetwork'
+export { TelepathChannelMock } from './TelepathChannelMock'
+export { SimpleStorageMock } from './SimpleStorageMock'
+export { InteractivePromise } from './InteractivePromise'
